requirements.txt
art==6.5
seqlog==0.4.2
tabulate==0.9.0
coloredlogs==15.0.1
pika==1.3.2
Click==8.2.1
requests==2.32.4
netaddr==1.3.0
netifaces==0.11.0
RPi.GPIO==0.7.1
adafruit-blinka==8.62.0
adafruit-circuitpython-busdevice==5.2.13
adafruit-circuitpython-mcp230xx==2.5.18
Adafruit-PlatformDetect==3.81.0
Adafruit-PureIO==1.1.11
